Changing the default values for compression.
This still needs investigation, d=1.0 shall be indecenable from reference but
this is not the case right now. Also, while this parameter drastically affect the image quality,
it seems not affecting file size.

A simple experiment shall be conducted to ensure libjxl is used the right way which seems not be
the fact with the current implementation.